+++ Modules/Microsoft365DSC/Modules/M365DSCDRGUtil.psm1
@@ -1436,10 +1436,21 @@ function ConvertFrom-IntunePolicyAssignment
         [Parameter(Mandatory = $true)]
         [Array]
         $Assignments,
+
         [Parameter()]
         [System.Boolean]
         $IncludeDeviceFilter = $true
     )
+
+    if ($null -eq $Script:IntuneAssignmentFilters)
+    {
+        $Script:IntuneAssignmentFilters = Get-MgBetaDeviceManagementAssignmentFilter -All -ErrorAction SilentlyContinue | ForEach-Object {
+            @{
+                FilterId = $_.Id
+                DisplayName = $_.DisplayName
+            }
+        }
+    }
 
     $assignmentResult = @()
     foreach ($assignment in $Assignments)
@@ -1507,7 +1518,9 @@ function ConvertFrom-IntunePolicyAssignment
             }
             if ($null -ne $assignment.Target.DeviceAndAppManagementAssignmentFilterId)
             {
-                $hashAssignment.Add('deviceAndAppManagementAssignmentFilterId', $assignment.Target.DeviceAndAppManagementAssignmentFilterId)
+                $filterId = $assignment.Target.DeviceAndAppManagementAssignmentFilterId
+                $hashAssignment.Add('deviceAndAppManagementAssignmentFilterId', $filterId)
+                $hashAssignment.Add('deviceAndAppManagementAssignmentFilterDisplayName', (($Script:IntuneAssignmentFilters | Where-Object -FilterScript { $_.FilterId -eq $filterId }).DisplayName))
             }
         }
 
@@ -1531,6 +1544,16 @@ function ConvertTo-IntunePolicyAssignment
         $IncludeDeviceFilter = $true
     )
 
+    if ($null -eq $Script:IntuneAssignmentFilters)
+    {
+        $Script:IntuneAssignmentFilters = Get-MgBetaDeviceManagementAssignmentFilter -All -ErrorAction SilentlyContinue | ForEach-Object {
+            @{
+                FilterId = $_.Id
+                DisplayName = $_.DisplayName
+            }
+        }
+    }
+
     if ($null -eq $Assignments)
     {
         return ,@()
@@ -1542,10 +1565,23 @@ function ConvertTo-IntunePolicyAssignment
         $target = @{"@odata.type" = $assignment.dataType}
         if ($IncludeDeviceFilter)
         {
-            if ($null -ne $assignment.DeviceAndAppManagementAssignmentFilterType)
+            if ($null -ne $assignment.DeviceAndAppManagementAssignmentFilterType -and $assignment.DeviceAndAppManagementAssignmentFilterType -ne 'none')
             {
-                $target.Add('deviceAndAppManagementAssignmentFilterType', $assignment.DeviceAndAppManagementAssignmentFilterType)
-                $target.Add('deviceAndAppManagementAssignmentFilterId', $assignment.DeviceAndAppManagementAssignmentFilterId)
+                $filter = $Script:IntuneAssignmentFilters | Where-Object -FilterScript { $_.FilterId -eq $assignment.DeviceAndAppManagementAssignmentFilterId }
+                if ($null -eq $filter)
+                {
+                    $filter = $Script:IntuneAssignmentFilters | Where-Object -FilterScript { $_.DisplayName -eq $assignment.DeviceAndAppManagementAssignmentFilterDisplayName }
+                    if ($null -eq $filter)
+                    {
+                        Write-Warning -Message "Assignment filter with DisplayName {$($assignment.DeviceAndAppManagementAssignmentFilterDisplayName)} not found in the directory. Please update your DSC resource extract with the correct filterId or filterDisplayName."
+                    }
+                }
+
+                if ($null -ne $filter)
+                {
+                    $target.Add('deviceAndAppManagementAssignmentFilterType', $assignment.DeviceAndAppManagementAssignmentFilterType)
+                    $target.Add('deviceAndAppManagementAssignmentFilterId', $filter.FilterId)
+                }
             }
         }
         if ($assignment.dataType -like '*CollectionAssignmentTarget')
@@ -1617,6 +1653,16 @@ function ConvertFrom-IntuneMobileAppAssignment
         $IncludeDeviceFilter = $true
     )
 
+    if ($null -eq $Script:IntuneAssignmentFilters)
+    {
+        $Script:IntuneAssignmentFilters = Get-MgBetaDeviceManagementAssignmentFilter -All -ErrorAction SilentlyContinue | ForEach-Object {
+            @{
+                FilterId = $_.Id
+                DisplayName = $_.DisplayName
+            }
+        }
+    }
+
     $assignmentResult = @()
     foreach ($assignment in $Assignments)
     {
@@ -1666,16 +1712,6 @@ function ConvertFrom-IntuneMobileAppAssignment
 
         $hashAssignment.Add('intent', $assignment.intent.ToString())
 
-        # $concatenatedSettings = $assignment.settings.ToString() -join ','
-        # $hashAssignment.Add('settings', $concatenatedSettings)
-        # $hashSettings = @{}
-        # foreach ($setting in $assignment.Settings)
-        # {
-        #   $hashSettings.Add('datatype', $setting.dataType)
-        #   $hashSettings.Add('uninstallOnDeviceRemoval', $setting.uninstallOnDeviceRemoval)
-        # }
-        # $hashAssignment.Add('settings', $hashSettings)
-
         if ($IncludeDeviceFilter)
         {
             if ($null -ne $assignment.Target.DeviceAndAppManagementAssignmentFilterType)
@@ -1684,7 +1720,9 @@ function ConvertFrom-IntuneMobileAppAssignment
             }
             if ($null -ne $assignment.Target.DeviceAndAppManagementAssignmentFilterId)
             {
-                $hashAssignment.Add('deviceAndAppManagementAssignmentFilterId', $assignment.Target.DeviceAndAppManagementAssignmentFilterId)
+                $filterId = $assignment.Target.DeviceAndAppManagementAssignmentFilterId
+                $hashAssignment.Add('deviceAndAppManagementAssignmentFilterId', $filterId)
+                $hashAssignment.Add('deviceAndAppManagementAssignmentFilterDisplayName', (($Script:IntuneAssignmentFilters | Where-Object -FilterScript { $_.FilterId -eq $filterId }).DisplayName))
             }
         }
 
@@ -1708,6 +1746,16 @@ function ConvertTo-IntuneMobileAppAssignment
         $IncludeDeviceFilter = $true
     )
 
+    if ($null -eq $Script:IntuneAssignmentFilters)
+    {
+        $Script:IntuneAssignmentFilters = Get-MgBetaDeviceManagementAssignmentFilter -All -ErrorAction SilentlyContinue | ForEach-Object {
+            @{
+                FilterId = $_.Id
+                DisplayName = $_.DisplayName
+            }
+        }
+    }
+
     if ($null -eq $Assignments)
     {
         return ,@()
@@ -1720,10 +1768,23 @@ function ConvertTo-IntuneMobileAppAssignment
         $target = @{"@odata.type" = $assignment.dataType}
         if ($IncludeDeviceFilter)
         {
-            if ($null -ne $assignment.DeviceAndAppManagementAssignmentFilterType)
+            if ($null -ne $assignment.DeviceAndAppManagementAssignmentFilterType -and $assignment.DeviceAndAppManagementAssignmentFilterType -ne 'none')
             {
-                $target.Add('deviceAndAppManagementAssignmentFilterType', $assignment.DeviceAndAppManagementAssignmentFilterType)
-                $target.Add('deviceAndAppManagementAssignmentFilterId', $assignment.DeviceAndAppManagementAssignmentFilterId)
+                $filter = $Script:IntuneAssignmentFilters | Where-Object -FilterScript { $_.FilterId -eq $assignment.DeviceAndAppManagementAssignmentFilterId }
+                if ($null -eq $filter)
+                {
+                    $filter = $Script:IntuneAssignmentFilters | Where-Object -FilterScript { $_.DisplayName -eq $assignment.DeviceAndAppManagementAssignmentFilterDisplayName }
+                    if ($null -eq $filter)
+                    {
+                        Write-Warning -Message "Assignment filter with DisplayName {$($assignment.DeviceAndAppManagementAssignmentFilterDisplayName)} not found in the directory. Please update your DSC resource extract with the correct filterId or filterDisplayName."
+                    }
+                }
+
+                if ($null -ne $filter)
+                {
+                    $target.Add('deviceAndAppManagementAssignmentFilterType', $assignment.DeviceAndAppManagementAssignmentFilterType)
+                    $target.Add('deviceAndAppManagementAssignmentFilterId', $filter.FilterId)
+                }
             }
         }
 
@@ -1845,6 +1906,12 @@ function Compare-M365DSCIntunePolicyAssignment
                     {
                         Write-Verbose 'FilterId specified, checking filterId'
                         $testResult = $assignment.deviceAndAppManagementAssignmentFilterId -eq $assignmentTarget.deviceAndAppManagementAssignmentFilterId
+
+                        if (-not $testResult)
+                        {
+                            Write-Verbose 'FilterId does not match, checking filterDisplayName'
+                            $testResult = $assignment.deviceAndAppManagementAssignmentFilterDisplayName -eq $assignmentTarget.deviceAndAppManagementAssignmentFilterDisplayName
+                        }
                     }
                     if (-not $testResult)
                     {